Join recent chat/group queries
